base UI complete(edit view)
