+++ Keep/Common/Extensions/View+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  View+Extension.swift
+//  Keep
+//
+//  Created by myung hoon on 23/03/2024.
+//
+
+import SwiftUI
+
+extension View {
+    func manualPopBack() -> some View {
+        ManualPopBackModifier(content: self)
+    }
+}
