+++ Keep/Features/Item/ItemInput/ItemInputItem.swift
@@ -1,5 +1,5 @@
 //
-//  ItemDetailViewItem.swift
+//  ItemInputItem.swift
 //  Keep
 //
 //  Created by myung hoon on 14/03/2024.
@@ -7,7 +7,8 @@
 
 import Foundation
 
-struct ItemDetailViewItem: Hashable {
+struct ItemInputItem: Hashable {
     let placeholder: String
     let inputType: ItemInputType
+    let displayType: ItemDisplayType
 }
