+++ Keep/Features/MainList/MainListItem.swift
@@ -1,0 +1,13 @@
+//
+//  MainListItem.swift
+//  Keep
+//
+//  Created by myung hoon on 23/03/2024.
+//
+
+import Foundation
+
+struct MainListItem: Identifiable, Hashable {
+    let id = UUID()
+    let title: String
+}
