+++ Keep/Features/MainList/MainListViewModel.swift
@@ -5,8 +5,26 @@
 //  Created by myung hoon on 29/02/2024.
 //
 
-import Foundation
+import Combine
 
 final class MainListViewModel: ObservableObject {
-    
+    let items = [MainListItem(title: "1"),
+                 MainListItem(title: "2"),
+                 MainListItem(title: "3"),
+                 MainListItem(title: "4"),
+                 MainListItem(title: "5"),
+                 MainListItem(title: "6"),
+                 MainListItem(title: "7"),
+                 MainListItem(title: "8"),
+                 MainListItem(title: "9"),
+                 MainListItem(title: "10"),
+                 MainListItem(title: "11"),
+                 MainListItem(title: "12"),
+                 MainListItem(title: "13"),
+                 MainListItem(title: "14"),
+                 MainListItem(title: "15"),
+                 MainListItem(title: "16"),
+                 MainListItem(title: "17"),
+                 MainListItem(title: "18")]
 }
+
